requirements.txt
flask
python-dotenv
flask_sqlalchemy
psycopg2-binary
flask-jwt-extended
python-docx
python-pptx
google-genai
psycopg2-binary
